% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{extract}
\alias{extract}
\title{Obtain only the top part of the NEWS file}
\usage{
extract(pkg, vpattern = "Changes in version")
}
\arguments{
\item{vpattern}{character(1) The 'grep' input for searching the versioning
line in the NEWS files. This usually starts with 'Changes in version'
but may differ.}
}
\description{
This function is used internally for collating the latest NEWS after
validation with the `validate` function.
}
