% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNEWSmd.R
\name{makeNEWSmd}
\alias{makeNEWSmd}
\title{Create a NEWS.md file from NEWS file}
\usage{
makeNEWSmd(newsfile = "inst/NEWS", overwrite = FALSE)
}
\arguments{
\item{newsfile}{character(1) The location of the NEWS file usually in
'inst/NEWS' (default)}

\item{overwite}{logical(1) Whether to overwrite the existing NEWS.md file
(default: FALSE)}
}
\value{
Saves a file with the name "NEWS.md" in the package directory
}
\description{
This function will add a package name line to the
current NEWS file and write a NEWS.md file for
viewing on GitHub. Future iterations will simplify
formatting changes to the NEWS file so that it can
be converted to a markdown file. Current files
should have markdown syntax in the NEWS file.
}
