% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgHeadings}
\alias{pkgHeadings}
\alias{headingsFormat}
\alias{headerCase}
\alias{updateHeading}
\title{Utilities for manipulating NEWS.md headers}
\usage{
pkgHeadings(packages, vpattern = "Changes in version")
}
\arguments{
\item{packages}{character() A vector of package names that correspond
to folders in the current directory.}

\item{vpattern}{character(1) The 'grep' input for searching the versioning
line in the NEWS files. This usually starts with 'Changes in version'
(default) but may differ.}

\item{pkg}{character(1) The name of a local package whose folder is in
in the current working directory.}

\item{dry.run}{logical(1) Whether to print changes to the console (TRUE) or
change the actual NEWS file in the package.}

\item{bumpHeaders}{logical(1) Whether to increase all markdown headers
by one (usually '##' to '###'; default TRUE)}

\item{from}{character(1) The Markdown header to identify in the NEWS.md file
(default '#' aka HTML tag 'h1')}

\item{to}{character(1) The Markdown header to replace in the NEWS.md file
(default '##' aka HTML tag 'h2')}
}
\description{
We provide several utilties for working with Markdown-style
NEWS.md files. Such operations include the manipulation of NEWS.md file
headers and header case.
}
\section{Headings}{

Obtain the NEWS file Markdown headings ('##') for a list of packages
with \code{pkgHeadings} which calls \code{headingsFormat} recursively.
It is useful to see whether there is consistency in the headings
corresponding to the 'Changes in version' line in the NEWS.md file

\code{headerCase} uses the \code{vpattern} input to replace any differently
formatted NEWS.md heading, e.g. \verb{CHANGES IN VERSION} to
\verb{Changes in version} where the differences are by case only.

\code{updateHeading} Increases the Markdown HTML h1 tags to h2 tags with the
option (\code{bumpHeaders = TRUE}) to bump all the headers for the remaining
sub-headers in order to maintain hierarchy, e.g., h2 to h3.
}

\examples{
\dontrun{
    pkgHeadings(
        c("MultiAssayExperiment", "curatedTCGAData", "TCGAutils",
        "cBioPortalData", "SingleCellMultiModal", "RTCGAToolbox",
        "RaggedExperiment")
    )

    headingsFormat("curatedTCGAData", "Changes in version")

    headerCase("RTCGAToolbox")

    updateHeading("cBioPortalData")
}

}
