% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsfeed-package.R
\docType{package}
\name{newsfeed-package}
\alias{newsfeed}
\alias{newsfeed-package}
\title{newsfeed: Fetch World Breaking News and Render HTML}
\description{
Fetches global breaking news feeds via C++ backend routines and exports the formatted output to an HTML report.
}
\author{
\strong{Maintainer}: Wayne Stewart \email{wayne.s.stewart@ou.edu}

Authors:
\itemize{
  \item Wayne Stewart \email{wayne.s.stewart@ou.edu}
}

}
\keyword{internal}
