% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{Convert a NEWS file to work with pkgdown}
\usage{
reformat(pkg = ".", backup = ".bak")
}
\arguments{
\item{pkg}{character(1) The directory location of the package usually the
current directory (default: ".")}

\item{backup}{character(1) The file extension to add for backing up the
original news file (default: '.bak')}
}
\value{
Saves a "NEWS.md" file with alternative format
}
\description{
This function will take an existing NEWS.md file and convert it to
pkgdown readable format. The former format should have headings that begin
with
\itemize{
    \item{CHANGES for 'CHANGES IN VERSION'}
    \item{'New' for 'New features'}
    \item{'Bug' for 'Bug fixes and minor improvements'}
}
These headings usually begin with 'h2' headers or double hashtag. An n-1
operation is performed on these headings (if the heading had two hashtags
'h2', they are now converted to one hashtag 'h1').
}
