% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameHeadings.R
\name{renameHeadings}
\alias{renameHeadings}
\title{Change the NEWS.md file headings}
\usage{
renameHeadings(pkg, backup = ".bak")
}
\arguments{
\item{pkg}{character(1) The directory location of the package usually the
current directory (default: ".")}

\item{backup}{character(1) The file extension to add for backing up the
original news file (default: '.bak')}
}
\value{
Saves a "NEWS.md" file with substituted headers
}
\description{
Some `NEWS.md` files have the package name in the heading. This is not
yet supported in the `collect` code. This convenience function will help
the user modify those headings to exclude the package name and include
the keywords `Changes in version` for aggregating news files across
packages. Version control of the NEWS files is highly recommended.
By default, the function saves a `*.bak` backup NEWS file in the package
directory.
}
\examples{

# create phony DESC for usethis
# otherwise temp 'package' is not recognized
desc <- usethis:::build_description(
    basename(tempdir()), roxygen = TRUE, fields = list()
)
lines <- desc$str(by_field = TRUE, normalize = FALSE, mode = "file")
usethis:::write_over(
    file.path(tempdir(), "DESCRIPTION"), lines, quiet = TRUE
)

tempNews <- file.path(tempdir(), "NEWS.md")
writeLines("# TempPackage 0.99.0", con = file(tempNews))
renameHeadings(pkg = tempdir())
readLines(tempNews)

}
