% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Check that NEWS and DESCRIPTION package versions are correct}
\usage{
validate(pkg, vpattern = "Changes in version")
}
\arguments{
\item{vpattern}{character(1) The 'grep' input for searching the versioning
line in the NEWS files. This usually starts with 'Changes in version'
but may differ.}
}
\description{
The NEWS file contains version information that can be parsed and
checked against the package version in the DESCRIPTION file.
}
\examples{
\dontrun{
validate("MultiAssayExperiment")
validate("RaggedExperiment")
}

}
