% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Compile NEWS files from several packages}
\usage{
collect(packages, vpattern = "Changes in version")
}
\arguments{
\item{packages}{A character vector of packages installed on the system}

\item{vpattern}{The 'grep' input for searching the versioning line in the
NEWS files. This usually starts with 'Changes in version' but may differ.}
}
\description{
This package will take the first chunk of a NEWS file and
return it as a character vector
}
\examples{

collect("newsfeed")

}
